--- modulo5/classes-encapsulamento/src/Bank.ts
@@ -1,9 +1,0 @@
-import { UserAccount } from './UserAccount'
-
-// export class Bank {
-//     private accounts: UserAccount[]
-
-//     constructor(accounts: UserAccount[] {
-//         this.accounts = accounts
-//     })
-// }
